chore: move stream from main player file
